Maintain seed data in a json file and read/insert from that

+++ capstone/src/main/java/com/google/step/data/Restaurant.java
@@ -20,7 +20,7 @@ import java.util.List;
 
 /** A restaurant to be displayed. */
 public final class Restaurant {
-  private final int restaurantKey;
+  private final long restaurantKey;
   private final String name;
   private final GeoPt location;
   private final String story;
@@ -30,7 +30,7 @@ public final class Restaurant {
   private final String status;
   private final double score;
 
-  public Restaurant(int restaurantKey, String name, GeoPt location, String story,
+  public Restaurant(long restaurantKey, String name, GeoPt location, String story,
       List<String> cuisine, String phone, String website, String status, double score) {
     this.restaurantKey = restaurantKey;
     this.name = name;
